Restructred the repo
