+++ Algorithms/Vowels.py
