+++ Comp/translator.py
