+++ Fun Experiments/anagram.py
