+++ Fun Experiments/char_distance.py
