+++ Fun Experiments/music.py
