+++ Fun Experiments/ps_check.py
